Created a interface for DB instructions
